fix(suite-native): scan for Bluetooth devices only when the app is active
